initial files: bring in sample apps directory modeled after Zephyr RTOS project.
The samples here will most often be borrowed from Zephyr's sample apps and
adapted for a particular board, and or modified to test some additional feature
not in the original sample.  If we are good we will include a README with each
sample with instructions for how to build the sample app, and a description of
expected output.