utils: add pair of functions to shell wrapper to openocd invocation.
Add a pair of shell functions which call openocd as closely as possible to
the way the recovery board v3.1's makefile calls openocd.

Note, flashing works and starting the gdb server via openocd appears to work
identically to the first pair of functions in the shell wrapper to openocd.
Those represent Zephyr's way of calling this flash and debugging utility.