app: add custom shell commands, add stub files for data keeping module.
The data "keeper" is designed to hold data gathered by other modules.  This
module is a culminator as well as a keeper of data.  The data needs to be
shared with logic which decides when to send certain messages and more
importantly, when to release a parachute drogue or main.

Also some intermediate comments added to adc.c.